% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_emotb.R
\name{apathy_adult}
\alias{apathy_adult}
\title{Score the NIH Toolbox Emotion Measures (2019)
This is for the raw score to t-score conversions for bank/fixed forms}
\usage{
apathy_adult(raw)
}
\arguments{
\item{df}{a data.frame containing the TB Emotion Measures items to be scored}

\item{drop_items}{whether to remove the item-level data from the \code{df}. Default: FALSE}
}
\description{
Score the NIH Toolbox Emotion Measures (2019)
This is for the raw score to t-score conversions for bank/fixed forms
}
\details{
Adds two columns, \code{_Theta} and \code{_Theta_T}, to \code{df} containing the nih toolbox emotion measures, respectively.

Note: scoring is based on the 2019 raw score to t-score conversion tables manual 

Note: some measures were omitted from the manual, in which case older versions were used

See: \url{http://www.healthmeasures.net/images/nihtoolbox/Technical_Manuals/Emotion/NIH_TB_Emotion_Raw_Score_to_T-Score_Conversion_Tables_Manual_3.19.19.pdf}
}
\author{
Michael Hallquist & Melanie Glatz
}
