% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_helpers.R
\name{check_singularity}
\alias{check_singularity}
\title{Probe singular fits of random effects in mixed effects regression models}
\usage{
check_singularity(lmerobj)
}
\arguments{
\item{m}{lmer model object}
}
\value{
Concerning RE estimates in m.
}
\description{
Probe singular fits of random effects in mixed effects regression models
}
\author{
Michael Hallquist
}
