% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convolve_fft_cpp}
\alias{convolve_fft_cpp}
\title{Internal function to convolve two vectors using FFT}
\arguments{
\item{a}{A vector}

\item{b}{A vector}
}
\value{
A vector containing the convolution of a and b
}
\description{
Internal function to convolve two vectors using FFT
}
\details{
This is an internal function
      Same result as, in R:
      convolve(a, b, conj=TRUE, type="open")
}
