% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_utility_fx.R
\name{convolve_regressor}
\alias{convolve_regressor}
\title{Helper function used by the summed runs and separate runs convolution steps}
\usage{
convolve_regressor(reg, n_vols, drop_volumes = 0, convolve = TRUE,
  normalization = "none", center_values = TRUE, convmax_1 = FALSE,
  high_pass = NULL, tr = NULL, hrf_parameters = c(a1 = 6, a2 = 12, b1 =
  0.9, b2 = 0.9, cc = 0.35))
}
\arguments{
\item{reg}{A matrix for a regressor containing onset, duration, and value on the columns}

\item{n_vols}{The number of volumes in the convolved output (before applying \code{drop_volumes})}

\item{drop_volumes}{The number of volumes to drop from the beginning of the regressor}

\item{convolve}{If \code{TRUE}, the regressor is convolved with the HRF. If \code{FALSE},
the regressor values are simply aligned onto the time grid without convolution
based on the corresponding onsets, durations, and values.}

\item{normalization}{The HRF normalization method used: "none", "durmax_1", or "evtmax_1"}

\item{convmax_1}{If \code{TRUE}, rescale convolved regressors to a maximum height of 1 to equilibrate scaling
across runs and subjects.}

\item{high_pass}{The cutoff frequency (in Hz) used for high-pass filtering. If \code{NULL}, no filtering is applied.}

\item{tr}{The repetition time (in seconds) for the fMRI scan.}
}
\description{
Helper function used by the summed runs and separate runs convolution steps
}
\author{
Michael Hallquist
}
\keyword{internal}
