% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_heatmap.R
\name{cor_heatmap}
\alias{cor_heatmap}
\title{Function to generate ggplot correlation heatmap}
\usage{
cor_heatmap(
  df,
  cormat = NULL,
  alphaSort = TRUE,
  base_size = 14,
  tileTextSize = 3.8,
  title = NULL
)
}
\arguments{
\item{df}{data.frame to plot}

\item{cormat}{Alternatively, pass a correlation matrix directory (not \code{df})}

\item{alphaSort}{Whether to sort variables in alphabetic order}

\item{base_size}{The base_size argument to the ggplot theme, controlling font sizes}

\item{tileTextSize}{The size of text inside the cells of the correlation matrix}

\item{title}{optional title added to ggplot}
}
\description{
Function to generate ggplot correlation heatmap
}
\examples{
  library(dplyr)
  data(iris)

  cor_heatmap(iris) #Species is not numeric and will be dropped automatically

}
