% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_with_target.R
\name{cor_with_target}
\alias{cor_with_target}
\title{This function correlates a set of target variables (target) with a set of other variables (withvars).
The general goal is to avoid sifting through gargantuan correlation matrices when only some cells are of interest a priori.}
\usage{
cor_with_target(
  df,
  omit = NULL,
  target,
  withvars = NULL,
  pmin = NULL,
  partial = NULL,
  absrmin = NULL,
  digits = 3,
  prewhiten = FALSE,
  orderbyr = FALSE
)
}
\arguments{
\item{df}{The data.frame containing data to be correlated}

\item{omit}{A vector of variables that should not be included in the \code{withvars}.}

\item{target}{A vector of variable names that are correlated with \code{withvars}.}

\item{withvars}{A vector of variable names, each of which is correlates with each \code{target}. If omitted, all variables are included.}

\item{pmin}{The highest correlation p-value that should be included in the output (e.g., to omit non-sig correlations)}

\item{partial}{A character vector of variable names to partial out of the target correlations}

\item{absrmin}{The minimum absolute r value to print}

\item{digits}{The number of digits to return for correlations and p values}

\item{prewhiten}{Whether to remove the temporal autocorrelation structure of each variable prior correlation (using ARIMA)}

\item{orderbyr}{Whether to sort the correlation output by the magnitude of the correlation}
}
\value{
A list of correlations where each element is a vector of a target variable with all withvvars
}
\description{
This function correlates a set of target variables (target) with a set of other variables (withvars).
The general goal is to avoid sifting through gargantuan correlation matrices when only some cells are of interest a priori.
}
\author{
Michael Hallquist
}
