% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve_funcs.R
\name{deconvolve_nlreg}
\alias{deconvolve_nlreg}
\title{R port of Bush and Cisler 2013, Magnetic Resonance Imaging
Adapted from the original provided by Keith Bush}
\usage{
deconvolve_nlreg(
  BOLDobs,
  kernel,
  nev_lr = 0.01,
  epsilon = 0.005,
  beta = 40,
  normalize = TRUE,
  trim_kernel = TRUE
)
}
\arguments{
\item{BOLDobs}{observed BOLD timeseries}

\item{kernel}{assumed kernel of the BOLD signal (e.g., from spm_hrf)}

\item{nev_lr}{learning rate for the assignment of neural events. Default: .01}

\item{epsilon}{relative error change (termination condition). Default: .005}

\item{beta}{slope of the sigmoid transfer function (higher = more nonlinear)}

\item{normalize}{whether to unit-normalize (z-score) \code{BOLDobs} before deconvolution. Default: TRUE}

\item{trim_kernel}{whether to remove the first K time points from the deconvolved vector, corresponding to
kernel leftovers from convolution. Default: TRUE}
}
\value{
A time series of the same length containing reconstructed neural events
}
\description{
R port of Bush and Cisler 2013, Magnetic Resonance Imaging
Adapted from the original provided by Keith Bush
}
\details{
This function deconvolves the BOLD signal using Bush 2011 method

Author:      Keith Bush, PhD
Institution: University of Arkansas at Little Rock
Date:        Aug. 9, 2013

The original code did not unit normalize the BOLD signal in advance, but in my testing, this
proves useful in many cases (unless you want to mess with the learning rate a lot), especially
when the time series has a non-zero mean (e.g., mean 100).
}
\author{
Keith Bush
}
