% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_nuisrun1}
\alias{example_nuisrun1}
\title{example nuisance regressors for run1 used in build_design_matrix}
\format{
An object of class \code{data.frame} with 300 rows and 2 columns.
}
\usage{
data(example_nuisrun1)
}
\description{
example nuisance regressors for run1 used in build_design_matrix
}
\keyword{datasets}
