% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_utility_fx.R
\name{fir1Bandpass}
\alias{fir1Bandpass}
\title{function to FIR-1 bandpass filter a signal. Can low- or high-pass filter by specifying 0 for low or >= Nyquist for high.}
\usage{
fir1Bandpass(x, TR = 2, low = 0.009, high = 0.08, n = 250,
  plotFilter = FALSE, forward_reverse = TRUE, padx = 0, detrend = 1)
}
\description{
function to FIR-1 bandpass filter a signal. Can low- or high-pass filter by specifying 0 for low or >= Nyquist for high.
}
\keyword{internal}
