% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_helpers.R
\name{flex_vif}
\alias{flex_vif}
\title{Render Variance Inflation Factor (VIF) in flextable}
\usage{
flex_vif(m)
}
\arguments{
\item{m}{model object}
}
\value{
VIF of predictors in model.
}
\description{
Render Variance Inflation Factor (VIF) in flextable
}
\author{
Michael Hallquist
}
