% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_utility_fx.R
\name{fmri.stimulus}
\alias{fmri.stimulus}
\title{Convolve a regressor with a hemodynamic response function for fMRI analysis.}
\usage{
fmri.stimulus(scans = 1, onsets = c(1), durations = c(1), values = c(1),
  center_values = FALSE, rm_zeros = TRUE, convolve = TRUE, rt = 3,
  times = NULL, demean = TRUE, a1 = 6, a2 = 12, b1 = 0.9, b2 = 0.9,
  cc = 0.35, parmax1 = FALSE)
}
\description{
Extended from \code{fmri} package to allow for continuous-valued regressor,
which is passed using the values parameter.
}
\details{
The function also supports mean centering of parametric regressor prior to convolution
to dissociate it from stimulus occurrence (when event regressor also in model)
}
