% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neuromap_med_list.R
\name{get_neuromap_med_list}
\alias{get_neuromap_med_list}
\title{Retrieve medication list for NeuroMAP}
\usage{
get_neuromap_med_list(
  med_info_path =
    "Studies/NeuroMAP/Data/Clinical/NMAP Clinical Interview Medication List.xlsx",
  survey_name = "NeuroMAP S3 - Pre-Session Survey",
  id_column = "Intro_ID",
  med_question = "Q11",
  common_dose_units = c("mg", "mcg", "pill", "pills", "tablet", "tablets", "ml", "grams",
    "g", "puff", "puffs")
)
}
\arguments{
\item{med_info_path}{The local path to the medication list Excel sheet containing a sheet called 'med_info'}

\item{survey_name}{The name of the survey that contains medication information in the Qualtrics Survey}

\item{id_column}{The name of the column in the Qualtrics form corresponding to the participant ID number}

\item{med_question}{A character vector of the name of the medication question in Qualtrics}

\item{common_dose_units}{A character vector of common dosage units that might appear in the Qualtrics form (keep lower case)}
}
\value{
A data.frame with the raw data and additional columns (medication class, dose amount, etc.)
}
\description{
Retrieve medication list for NeuroMAP
}
\author{
Zach Vig
}
