% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_neuromap_self_reports.R
\name{import_neuromap_self_reports}
\alias{import_neuromap_self_reports}
\title{Imports NeuroMAP Self Report Data from Qualtrics}
\usage{
import_neuromap_self_reports(
  info = FALSE,
  stats = FALSE,
  survey_name = "NeuroMAP S2 - Self Report",
  scales = "all",
  include_id = TRUE,
  include_dem = FALSE,
  path = getwd(),
  file_suffix = "",
  file_date = FALSE,
  add_to_envr = FALSE
)
}
\arguments{
\item{info}{Logical. If \code{TRUE}, displays information about the function and the scales that can be extracted.
Additionally, it assigns the global variable \code{scales_to_score} with the names of all the scales available for scoring.
Default is \code{FALSE}.}

\item{stats}{Logical. If \code{TRUE}, displays basic statistics about the self report data, such as the number of entries,
the number of completed entries, and the date of the latest entry. Default is \code{FALSE}.}

\item{survey_name}{Character. The name of the survey in Qualtrics. Default is "NeuroMAP S2 - Self Report".}

\item{scales}{Character or vector of characters. The scales to extract from the self report data. Available scales are:
"iip90" (Inventory for Interpersonal Problems), "ctq" (Childhood Trauma Questionnaire), "panas"
(Positive and Negative Affect Schedule), "pid5" (Personality Inventory for DSM-5), "asr" (Adult Self Report),
"bpq" (Borderline Personality Questionnaire), "fs" (Flourishing Scale), "dusi" (Short Form Revised Drug Use
Screening Inventory), "isc" (Interpersonal Stressors Circumplex), "upps" (Impulsive Behaviors Scale),
"bfi" (Big Five Inventory-2), "cts" (Conflict Tactics Scale), and "emotb" (Emotional Trauma Outcome Measures Battery).
Default is "all" to include all scales.}

\item{include_id}{Logical. If \code{TRUE}, includes the participant ID in the output files. Default is \code{TRUE}.}

\item{include_dem}{Logical. If \code{TRUE}, includes demographic information (e.g., sex, age) in the output files.
Default is \code{FALSE}.}

\item{path}{Character. The path where the output files will be saved. If the specified path does not exist, the function
will create the directories recursively. Default is the current working directory.}

\item{file_suffix}{Character. A suffix to add to the output file names. Default is an empty string.}

\item{file_date}{Logical. If \code{TRUE}, appends a timestamp with the format "_%b_%d_%I_%M_%p" to the output file names.
Default is \code{FALSE}.}

\item{add_to_envr}{Logical. If \code{TRUE}, returns a list of requested data.frames. Default is \code{FALSE}}
}
\description{
This function imports self report data for the NeuroMAP Study from Qualtrics and saves it as separate CSV files based on the specified scales.
}
\examples{
import_neuromap_self_reports(info = TRUE)
import_neuromap_self_reports(scales = c("ctq", "asr"), include_dem = TRUE)
import_neuromap_self_reports(stats = TRUE)
import_neuromap_self_reports(path = "~/my_output_files")
import_neuromap_self_reports(file_suffix = "_v2", file_date = TRUE)
}
\author{
Zach Vig
}
