% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_score_neuromap_sidp.R
\name{import_score_neuromap_sidp}
\alias{import_score_neuromap_sidp}
\title{Import and Score NeuroMAP SIDP Survey Data}
\usage{
import_score_neuromap_sidp(...)
}
\arguments{
\item{...}{Additional arguments passed to `qualtRics::fetch_survey`.}
}
\value{
A list containing two elements:
        - `summary`: a dataframe with the summary scores for each personality disorder.
        - `items`: a list of dataframes, each containing item-level data for a specific disorder.
}
\description{
This function imports survey data from the NeuroMAP S1 - SIDP, processes it,
and computes scores for different personality disorders based on survey responses.
It performs data cleaning, reformatting, column sorting, and calculation of both
categorical and dimensional scores for various personality disorders.
Additionally, it handles specific diagnostic criteria for certain disorders.
}
\examples{
# Example usage:
\dontrun{
result <- import_score_neuromap_sidp()
summary_data <- result$summary
item_data <- result$items
}

}
\author{
Nate Hall
}
