% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_project.R
\name{initialize_project}
\alias{initialize_project}
\title{create a new data analysis project}
\usage{
initialize_project(
  path,
  sub_dirs = c("analysis_scripts", "preproc_scripts", "support_fx", "data_raw",
    "data_preproc", "figures", "results", "reports"),
  git_setup = TRUE,
  open = TRUE,
  readme_text = paste0("Welcome to ", basename(path), "!")
)
}
\arguments{
\item{path}{full path of the project directory to be created}

\item{sub_dirs}{character vector of high-level directories to add. Defaults to DEPENd Lab standard practice.}

\item{git_setup}{logical. intitialize new git repository and perform initial commit}

\item{open}{logical. open newly Rstudio project}

\item{description}{text to display in README.md}
}
\description{
automatically creates project directory and creates .Rproj file
}
\details{
make sure to install github CLI to use gh commands: https://github.com/cli/cli#installation. Flesh this out at a later date
}
\examples{
\dontrun{
 initialize_project("~/Documents/github_repos/dimt_analysis")
 initialize_project("~/Documents/github_repos/dimt_analysis", git_setup = FALSE) # just creates directory structure with no git integration
}

}
\author{
Nate Hall
}
