% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installMissingPkgs.R
\name{installMissingPkgs}
\alias{installMissingPkgs}
\title{Installs packages that are missing based off a list provided. Usually, this list comes from a previous save of the R environment. 
This is for the 32-item version}
\usage{
installMissingPkgs()
}
\description{
Installs packages that are missing based off a list provided. Usually, this list comes from a previous save of the R environment. 
This is for the 32-item version
}
\details{
Note: this function is useful when updating R, as it allows you to re-install packages that were lost when transitioning from an older version to a newer version. 
will yield the 32 items from the test.
}
\author{
Michael Hallquist
}
