% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_impute_items.R
\name{mean_impute_items}
\alias{mean_impute_items}
\title{Function to impute the mean for missing items on a per-case (row) basis
if the proportion of missingness falls below a specified threshold (e.g. 0.2)}
\usage{
mean_impute_items(scale_df, scale_items = NULL, thresh = 0.2)
}
\arguments{
\item{scale_df}{a data.frame object whose columns are the items to be imputed}

\item{scale_items}{a character vector of items names the define the scale/subscale to be imputed}

\item{thresh}{the proportion of missingness [0..1) or number [1..] of missing values below which the mean will be imputed for missing items}
}
\description{
Function to impute the mean for missing items on a per-case (row) basis
if the proportion of missingness falls below a specified threshold (e.g. 0.2)
}
\details{
If the proportion of missingness is greater than \code{thresh} then no items will be imputed
and the observation will not be altered.
}
\author{
Michael Hallquist
}
