% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_helpers.R
\name{render_aictab}
\alias{render_aictab}
\title{Render a pretty AIC comparison table in flextable for iterative model building and manuscrips.}
\usage{
render_aictab(minfo, mobs = NULL, ...)
}
\arguments{
\item{minfo}{named character vector containing model infomation to print to the table. (e.g. fixed and random effects)}

\item{mobs}{list of object models to include in table}
}
\value{
A time series of the same length containing reconstructed neural events
}
\description{
Render a pretty AIC comparison table in flextable for iterative model building and manuscrips.
}
\author{
Michael Hallquist

Nate Hall
}
