% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_helpers.R
\name{render_flex_aictab}
\alias{render_flex_aictab}
\title{Render a pretty AIC comparison table in flextable for iterative model building and manuscripts.}
\usage{
render_flex_aictab(minfo, ..., mlist = NULL)
}
\arguments{
\item{minfo}{named character vector containing model infomation to print to the table. (e.g. fixed and random effects)}

\item{mlist}{list of object models to include in table}
}
\value{
publication-ready table of AIC comparisons in flextable format.
}
\description{
Render a pretty AIC comparison table in flextable for iterative model building and manuscripts.
}
\author{
Michael Hallquist

Nate Hall
}
