% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_onedrive.R
\name{retrieve_onedrive}
\alias{retrieve_onedrive}
\title{retrieve .csv and excel (xls/xlsx) files from OneDrive/Sharepoint site}
\usage{
retrieve_onedrive(item_path, sheet = NULL, spn = 1, rm_file = TRUE)
}
\arguments{
\item{item_path}{path to desired OneDrive sharepoint file. Treat the base directory as root.}

\item{sheet}{String of sheet to extract (if excel file). Defaults to NULL}

\item{spn}{Share point number. Can determine by running Microsoft365R::list_sharepoint_sites() and determining which site number is appropriate. Defaults to 1. If you have more than one sharepoint site on your account you may need to adjust this!}

\item{rm_file}{the function needs to download a copy of the file into your current directory. Leaving this set to TRUE (default) will remove the downloaded file.}
}
\value{
Extracted OneDrive file
}
\description{
retrieve .csv and excel (xls/xlsx) files from OneDrive/Sharepoint site
}
\examples{

\dontrun{
     master <- retrieve_onedrive("Studies/NeuroMAP/Participant Management/Master Spreadsheets/NeuroMAP Master UNC.xlsx") # returns all sheets in a named list
     tracker <- retrieve_onedrive("Studies/NeuroMAP/Participant Management/Master Spreadsheets/NeuroMAP Master UNC.xlsx", sheet = "Tracker") # extracts only the Tracker sheet
}

}
\author{
Nate Hall
}
