% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFSLCommand.R
\name{runFSLCommand}
\alias{runFSLCommand}
\title{Wrapper for running an fsl command safely within R}
\usage{
runFSLCommand(args, fsldir = NULL, stdout = NULL, stderr = NULL,
  echo = TRUE)
}
\arguments{
\item{args}{FSL command string to be run}

\item{fsldir}{Location of FSL installation. If NULL, the function}

\item{stdout}{File target for redirecting stdout. If NULL, stdout will not be captured}

\item{stderr}{File target for redirecting stderr. If NULL, stderr will not be captured}

\item{echo}{Whether to print FSL command to the screen}
}
\value{
The exit status of the executed FSL command. 0 for success, non-zero for failure
}
\description{
Wrapper for running an fsl command safely within R
}
\details{
This command ensures that $FSLDIR/bin is in the PATH and sources $FSLDIR/etc/fslconf/fsl.sh

I should probably just use the \code{fslr} package, but didn't know about it prior to writing this...
}
\examples{

\dontrun{
runFSLCommand("fslmaths inp_nif -mul 10 -abs -thr 15 manip_out")
}
}
\author{
Michael Hallquist
}
