% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_bpq.R
\name{score_bpq}
\alias{score_bpq}
\title{Score the Borderline Personality Questionnaire (BPQ)}
\usage{
score_bpq(
  df,
  item_prefix = "BPQ",
  max_impute = 0.2,
  drop_items = FALSE,
  keep_reverse_codes = FALSE,
  input_codes = c(1, 2),
  min_value = 0,
  max_value = 1,
  bad_items = NULL,
  add_alphas = TRUE
)
}
\arguments{
\item{df}{a data.frame containing the 80 BPQ items to be scored}

\item{item_prefix}{a character prefix of the items names in \code{df} to be scored. Default: "BPQ"}

\item{max_impute}{the proportion of missingness [0..1) or number [1..n] of missing values per scale.
Below this threshold, the person subscale mean will be imputed for missing items.}

\item{drop_items}{whether to remove the item-level data from the \code{df}. Default: FALSE}

\item{keep_reverse_codes}{whether to retain the reverse coded items (suffix "r")}

\item{input_codes}{the original T/F coding scheme of the input dataframe. These are converted to the specified min and max value.}

\item{min_value}{the minimum value for the item anchors, used in reverse scoring. Default: 0}

\item{max_value}{the highest value for the item anchors, used in reverse scoring. Default: 1}

\item{bad_items}{numeric value or vector of the items that need to be dropped before imputation or calculation of subscales}

\item{add_alphas}{whether to compute coefficient alpha for subscales and return a column attribute. Default: TRUE}
}
\description{
Score the Borderline Personality Questionnaire (BPQ)
}
\details{
Adds nine columns, \code{BPQ_impulsivity}, \code{BPQ_instability}, \code{BPQ_abandonment}, \code{BPQ_relationships}, \code{BPQ_selfimage}, \code{BPQ_suicide}, \code{BPQ_emptiness}, \code{BPQ_anger}, \code{BPQ_psychotic}, and \code{BPQ_total}, to \code{df} containing
the different subscales, respectively.

Note: the default BPQ scoring uses the mean of the items for the scales.

Note: the code assumes that pasting together the \code{item_prefix} and the numbers 1:80
will yield the 80 items from the test.
}
\author{
Michael Hallquist
}
