% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_cts.R
\name{score_cts}
\alias{score_cts}
\title{Score the Conflict Tactics Scale (CTS)}
\usage{
score_cts(df, item_prefix = "CTS", max_impute = 0.2,
  drop_items = FALSE, min_value = 1, max_value = 7)
}
\arguments{
\item{df}{a data.frame containing the 40 CTS items to be scored}

\item{item_prefix}{a character prefix of the items names in \code{df} to be scored. Default: "CTS"}

\item{max_impute}{the proportion of missingness [0..1) or number [1..n] of missing values per subscale.
Below this, the mean will be imputed for missing items.}

\item{drop_items}{whether to remove the item-level data from the \code{df}. Default: FALSE}

\item{min_value}{the lowest value for the items anchors, used to check response validity. Default: 1}

\item{max_value}{the highest value for the items anchors, used to check response validity. Default: 7}
}
\description{
Score the Conflict Tactics Scale (CTS)
}
\details{
UPDATE ME!
Adds two columns, \code{ECR_anxiety} and \code{ECR_avoidance}, to \code{df} containing
the anxiety and avoidance scales, respectively.

Note: the default CTS scoring uses the mean of the items for the scales.

Note: the code assumes that pasting together the \code{item_prefix} and the numbers 1:40
will yield the 40 items from the test.
}
\author{
Michael Hallquist
}
