% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_das.R
\name{score_das}
\alias{score_das}
\title{Scores the Dyadic Adjustment Scale (Spanier 1976)
This is for the 32-item version}
\usage{
score_das(
  df,
  item_prefix = "DAS",
  max_impute = 0.2,
  drop_items = FALSE,
  max_value = 7
)
}
\arguments{
\item{df}{a data.frame containing the 32 DAS items to be scored}

\item{item_prefix}{a character prefix of the items names in \code{df} to be scored. Default: "ECR"}

\item{max_impute}{the proportion of missingness [0..1) or number [1..] of missing values per scale. Below this, the mean will be imputed for missing items}

\item{drop_items}{whether to remove the item-level data from the \code{df}. Default: FALSE}

\item{max_value}{the highest value for the items anchors, used in reverse scoring. Default: 7}
}
\description{
Scores the Dyadic Adjustment Scale (Spanier 1976)
This is for the 32-item version
}
\details{
Adds five columns, \code{DASSat},  \code{DASCon}, \code{DASCoh}, \code{DASAffExp}, and \code{DASTotal} to \code{df} containing
the different subscaled.

Note: the default DAS scoring uses the mean of the items for the scales.

Note: the code assumes that pasting together the \code{item_prefix} and the numbers 1:32
will yield the 32 items from the test.
}
\author{
Michael Hallquist
}
