% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_ders.R
\name{score_ders}
\alias{score_ders}
\title{Score the Difficulties in Emotion Regulation Scale (DERS) (Gratz & Roemer, 2004)}
\usage{
score_ders(
  df,
  item_prefix = "DERS",
  max_impute = 0.2,
  drop_items = FALSE,
  keep_reverse_codes = FALSE,
  min_value = 1,
  max_value = 5
)
}
\arguments{
\item{df}{a data.frame containing the 36 DERS items to be scored}

\item{item_prefix}{a character prefix of the items names in \code{df} to be scored. Default: "DERS"}

\item{max_impute}{the proportion of missingness [0..1) or number [1..n] of missing values per scale.
Below this threshold, the person subscale mean will be imputed for missing items.}

\item{drop_items}{whether to remove the item-level data from the \code{df}. Default: FALSE}

\item{keep_reverse_codes}{whether to retain the reverse coded items (suffix "r")}

\item{min_value}{the minimum value for the item anchors, used in reverse scoring. Default: 1}

\item{max_value}{the highest value for the item anchors, used in reverse scoring. Default: 5}
}
\description{
Score the Difficulties in Emotion Regulation Scale (DERS) (Gratz & Roemer, 2004)
}
\details{
Adds seven columns, \code{DERS_nonaccept}, \code{DERS_goals}, \code{DERS_impulse}, \code{DERS_aware}, \code{DERS_strategies}, \code{DERS_clarity}, and \code{DERS_total}, to \code{df} containing
the different subscales, respectively.

Note: the default DERS scoring uses the mean of the items for the scales.

Note: the code assumes that pasting together the \code{item_prefix} and the numbers 1:36
will yield the 36 items from the test.
}
\author{
Michael Hallquist
}
