% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_dusi.R
\name{score_dusi}
\alias{score_dusi}
\title{Score the Short Version of the Revised Drug Use Screening Inventory (DUSI-R) (Tarter, 1990)}
\usage{
score_dusi(
  df,
  item_prefix = "DUSI_",
  max_impute = 0.2,
  drop_items = FALSE,
  input_codes = c(0, 1),
  false_value = 0,
  true_value = 1,
  bad_items = NULL,
  add_alphas = TRUE
)
}
\arguments{
\item{df}{a data.frame containing the 15 DUSI-R SF items to be scored}

\item{item_prefix}{a character prefix of the items names in \code{df} to be scored. Default: "DUSI_"}

\item{max_impute}{the proportion of missingness [0..1) or number [1..n] of missing values per scale. Default: 0.2
Below this threshold, the person subscale mean will be imputed for missing items.}

\item{drop_items}{whether to remove the item-level data from the \code{df}. Default: FALSE}

\item{input_codes}{the original T/F coding scheme of the input data.frame in the form c(False Code, True Code).
These are converted to the specified false and true value.}

\item{false_value}{the 'False' value for the item anchors, used in score validation. Default: 0}

\item{true_value}{the 'True' value for the item anchors, used in score validation. Default: 1}

\item{bad_items}{numeric value or vector of the items that need to be dropped before imputation or calculation of subscales}

\item{add_alphas}{whether to compute coefficient alpha for the total scale and return a column attribute. Default: TRUE}
}
\description{
Score the Short Version of the Revised Drug Use Screening Inventory (DUSI-R) (Tarter, 1990)
}
\details{
Adds two columns, \code{DUSI_total} and \code{DUSI_density}, to \code{df} containing the sum of scores and
mean of the items for the scales (i.e., the "absolute problem density"), respectively.

Note: the code assumes that pasting together the \code{item_prefix} and the numbers 1:15
will yield the 15 items from the test.
}
\author{
Michael Hallquist, Zach Vig
}
