% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_ecr.R
\name{score_ecr}
\alias{score_ecr}
\title{Score the experiences in close relationships -- revised (ECR-R) scale
This is for the 36-item version.}
\usage{
score_ecr(df, item_prefix = "ECR", max_impute = 0.2,
  drop_items = FALSE, keep_reverse_codes = FALSE, max_value = 7)
}
\arguments{
\item{df}{a data.frame containing the 36 ECR items to be scored}

\item{item_prefix}{a character prefix of the items names in \code{df} to be scored. Default: "ECR"}

\item{max_impute}{the proportion of missingness [0..1) or number [1..] of missing values per scale. Below this, the mean will be imputed for missing items}

\item{drop_items}{whether to remove the item-level data from the \code{df}. Default: FALSE}

\item{keep_reverse_codes}{whether to retain the reverse coded items (suffix "r")}

\item{max_value}{the highest value for the items anchors, used in reverse scoring. Default: 7}
}
\description{
Score the experiences in close relationships -- revised (ECR-R) scale
This is for the 36-item version.
}
\details{
Adds two columns, \code{ECR_anxiety} and \code{ECR_avoidance}, to \code{df} containing
the anxiety and avoidance scales, respectively.

Note: the default ECR scoring uses the mean of the items for the scales.

Note: the code assumes that pasting together the \code{item_prefix} and the numbers 1:36
will yield the 36 items from the test.
}
\author{
Michael Hallquist
}
