% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_emotb.R
\name{score_emotb}
\alias{score_emotb}
\title{Score the NIH Toolbox Emotion Measures (2019)}
\usage{
score_emotb(df, drop_items = FALSE)
}
\arguments{
\item{df}{The input data frame containing item responses.}

\item{drop_items}{Logical indicating whether to drop item-level data from the output. Default is FALSE.}
}
\value{
The modified data frame with EMOTB scores.
}
\description{
This function calculates EMOTB scores based on the provided data frame and parameters.
}
\details{
This function adds two columns, `_Theta` and `_Theta_T`, to `df` containing the NIH Toolbox emotion measures, respectively.
}
\note{
Scoring is based on the 2019 raw score to t-score conversion tables manual.

Some measures were omitted from the manual, in which case older versions were used.

The following age group definitions are used:
- TB_measure_adult: 18 years and older
- TB_measure_teen: 13-17 years and 8-17 years
- TB_measure_child: 8-12 years
}
\examples{
df <- data.frame(
  Age = c(20, 15, 25),
  Apathy_Raw = c(1, 2, 3),
  PercStrs_Raw = c(2, 1, 3),
  Self_Raw = c(3, 2, 1)
)
score_emotb(df)

}
\seealso{
For T-score conversion tables, see the [NIH TB Emotion Raw Score to T-Score Conversion Tables Manual](http://www.healthmeasures.net/images/nihtoolbox/Technical_Manuals/Emotion/NIH_TB_Emotion_Raw_Score_to_T-Score_Conversion_Tables_Manual_3.19.19.pdf).
}
\author{
Michael Hallquist, Melanie Glatz
}
