% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_isc.R
\name{score_isc}
\alias{score_isc}
\title{Score the Interpersonal Stressors Circumplex (ISC)}
\usage{
score_isc(df, item_prefix = "ISC", max_impute = 0.2,
  drop_items = FALSE, min_value = 1, max_value = 8)
}
\arguments{
\item{df}{a data.frame containing the 64 ISC items to be scored}

\item{item_prefix}{a character prefix of the items names in \code{df} to be scored. Default: "ISC"}

\item{max_impute}{the proportion of missingness [0..1) or number [1..n] of missing values per scale.
Below this threshold, the person subscale mean will be imputed for missing items.}

\item{drop_items}{whether to remove the item-level data from the \code{df}. Default: FALSE}

\item{min_value}{the minimum value for the item anchors. Default: 1}

\item{max_value}{the highest value for the item anchors. Default: 8}
}
\description{
Score the Interpersonal Stressors Circumplex (ISC)
}
\details{
Adds eight columns, \code{ISC_c1}, \code{ISC_c2}, \code{ISC_c3}, \code{ISC_c4}, \code{ISC_c5}, \code{ISC_c6}, \code{ISC_c7}, and \code{ISC_c8}, to \code{df} containing
the different subscales, respectively.

Note: the default ISC scoring uses the mean of the items for the scales.

Note: the code assumes that pasting together the \code{item_prefix} and the numbers 1:64
will yield the 64 items from the test.
}
\author{
Michael Hallquist
}
