% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_neo.R
\name{score_neo}
\alias{score_neo}
\title{Score the NEO Five-Factor Inventory (NEO-FFI) scale (Costa & McCrae, 1992)
This is for the 60-item version}
\usage{
score_neo(df, item_prefix = "NEO", max_impute = 0.2,
  drop_items = FALSE, keep_reverse_codes = FALSE, min_value = 1,
  max_value = 5)
}
\arguments{
\item{df}{a data.frame containing the 60 NEO items to be scored}

\item{item_prefix}{a character prefix of the items names in \code{df} to be scored. Default: "NEO"}

\item{max_impute}{the proportion of missingness [0..1) or number [1..n] of missing values per scale.
Below this threshold, the person subscale mean will be imputed for missing items.}

\item{drop_items}{whether to remove the item-level data from the \code{df}. Default: FALSE}

\item{keep_reverse_codes}{whether to retain the reverse coded items (suffix "r")}

\item{min_value}{the minimum value for the item anchors, used in reverse scoring. Default: 1}

\item{max_value}{the highest value for the item anchors, used in reverse scoring. Default: 5}
}
\description{
Score the NEO Five-Factor Inventory (NEO-FFI) scale (Costa & McCrae, 1992)
This is for the 60-item version
}
\details{
Adds five columns, \code{NEO_neurot}, \code{NEO_extra}, \code{NEO_open}, \code{NEO_agree}, and \code{NEO_cons}, to \code{df} containing 
the NEO-FFI scales, respectively.

Note: the default NEO scoring uses the mean of the items for the scales.

Note: the code assumes that pasting together the \code{item_prefix} and the numbers 1:60
will yield the 60 items from the test.
}
\author{
Michael Hallquist
}
