% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_paibor.R
\name{score_paibor}
\alias{score_paibor}
\title{Score PAI-BOR subscales for affective instability, indentity problems, negative relationships, and self-harm}
\usage{
score_paibor(
  df,
  item_prefix = "PAIBOR_",
  max_impute = 0.2,
  min_value = 0,
  max_value = 3
)
}
\arguments{
\item{df}{dataframe containing the PAI-BOR items to be scored}

\item{item_prefix}{a character prefix of the items names in \code{df} to be scored. Default: "PAIBOR_"}

\item{max_impute}{the proportion of missingness [0..1) or number [1..n] of missing values per scale.
Below this threshold, the person subscale mean will be imputed for missing items.}

\item{min_value}{the minimum value for the item anchors, used in reverse scoring. Default: 0}

\item{max_value}{the highest value for the item anchors, used in reverse scoring. Default: 3}
}
\value{
A data frame with the PAI-BOR subscales scores and total PAI-BOR score
}
\description{
This function scores the PAI-BOR subscales for all subject IDs in the tracker tab of the master spreadsheet.
}
\author{
Nidhi Desai
}
