% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_panas.R
\name{score_panas}
\alias{score_panas}
\title{Score the Positive and Negative Affect Schedule (PANAS) scale}
\usage{
score_panas(
  df,
  item_prefix = "PANAS",
  max_impute = 0.2,
  drop_items = FALSE,
  min_value = 1,
  max_value = 5
)
}
\arguments{
\item{df}{a data.frame containing the 20 PANAS items to be scored}

\item{item_prefix}{a character prefix of the items names in \code{df} to be scored. Default: "PANAS"}

\item{max_impute}{the proportion of missingness [0..1) or number [1..n] of missing values per scale.
Below this threshold, the person subscale mean will be imputed for missing items.}

\item{drop_items}{whether to remove the item-level data from the \code{df}. Default: FALSE}

\item{min_value}{the minimum value for the item anchors. Default: 1}

\item{max_value}{the highest value for the item anchors. Default: 5}
}
\description{
Score the Positive and Negative Affect Schedule (PANAS) scale
}
\details{
Adds two columns, \code{PANAS_positive} and \code{PANAS_negative}, to \code{df} containing
the positive and negative affect scales, respectively.

Note: the default PANAS scoring uses the mean of the items for the scales.

Note: the code assumes that pasting together the \code{item_prefix} and the numbers 1:20
will yield the 20 items from the test.
}
\author{
Michael Hallquist
}
