% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_pid5.R
\name{score_pid5}
\alias{score_pid5}
\title{Score the Personality Inventory for DSM-5 Faceted Brief Form -- 100 items PID-5-BF (Maples et al., 2013)}
\usage{
score_pid5(
  df,
  item_prefix = "PID_",
  max_impute = 0.2,
  drop_items = FALSE,
  bad_items = NULL,
  min_value = 0,
  max_value = 3,
  add_alphas = TRUE
)
}
\arguments{
\item{df}{a data.frame containing the 100 PID-5-BF items to be scored}

\item{item_prefix}{a character prefix of the items names in \code{df} to be scored. Default: "PID"}

\item{max_impute}{the proportion of missingness [0..1) or number [1..n] of missing values per scale.
Below this threshold, the person subscale mean will be imputed for missing items.}

\item{drop_items}{whether to remove the item-level data from the \code{df}. Default: FALSE}

\item{bad_items}{numeric value or vector of the items that need to be dropped before imputation or calculation of subscales}

\item{min_value}{the minimum value for the item anchors, used in reverse scoring. Default: 0}

\item{max_value}{the highest value for the item anchors, used in reverse scoring. Default: 3}

\item{add_alphas}{whether to compute coefficient alpha for the total scale and return a column attribute. Default: TRUE}
}
\description{
Score the Personality Inventory for DSM-5 Faceted Brief Form -- 100 items PID-5-BF (Maples et al., 2013)
}
\details{
Adds thirty columns to \code{df} containing the different trait facet and trait domain subscales, respectively.

Note: the default PID scoring uses the mean of the items for the scales.

Note: the code assumes that pasting together the \code{item_prefix} and the numbers 1:100
will yield the 100 items from the test.
}
\author{
Michael Hallquist
}
