% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_rist.R
\name{score_rist}
\alias{score_rist}
\title{Calculate RIST T-scores, RIST index, percentile and 90 percent confidence interval}
\usage{
score_rist(
  qualtrics_api_key = "",
  qualtrics_base_url = "unc.sjc1.qualtrics.com",
  qualtrics_surveyID = "SV_1YvGg9Emrw1cE3Y",
  return_dates = TRUE
)
}
\arguments{
\item{qualtrics_api_key}{Important credentials you need to authenticate with the Qualtrics API. You only need to provide this if your Qualtrics credentials are not already loaded.}

\item{qualtrics_base_url}{Important credentials you need to authenticate with the Qualtrics API. You only need to provide this if your Qualtrics credentials are not already loaded.}

\item{qualtrics_surveyID}{surveyID of the Qualtrics survey you are trying to pull data from. Default set to NeuroMAP RIST survey.}

\item{return_dates}{Set to TRUE if you would like to get the date of RIST session, date of birth, and subject age in the output dataframe. Default: TRUE.}
}
\description{
This function scores the RIST Odd-item-out items and Guess-What for all subject IDs in the RIST Qualtrics survey.
}
\examples{
score_rist() # Qualtrics API key already setup and using other default input parameters tuned for NeuroMAP RIST Qualtrics survey
score_rist(qualtrics_api_key = "cLCPP23vnvkbvkEltXRJCtTttttL13TZswEZZ9Zq") # use your qualtrics API key
score_rist(qualtrics_surveyID = "DR_1YvGg9Emrw1cE3Y") # use a different Qualtrics survey than NeuroMAP RIST
}
\author{
Nidhi Desai
}
