% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_rse.R
\name{score_rse}
\alias{score_rse}
\title{Score the Relationship and Sex Education (RSE) scale}
\usage{
score_rse(df, item_prefix = "RSE", max_impute = 0.2,
  drop_items = FALSE, keep_reverse_codes = FALSE, min_value = 1,
  max_value = 4)
}
\arguments{
\item{df}{a data.frame containing the 10 RSE items to be scored}

\item{item_prefix}{a character prefix of the items names in \code{df} to be scored. Default: "RSE"}

\item{max_impute}{the proportion of missingness [0..1) or number [1..n] of missing values per scale.
Below this threshold, the person subscale mean will be imputed for missing items.}

\item{drop_items}{whether to remove the item-level data from the \code{df}. Default: FALSE}

\item{keep_reverse_codes}{whether to retain the reverse coded items (suffix "r")}

\item{min_value}{the minimum value for the item anchors, used in reverse scoring. Default: 1}

\item{max_value}{the highest value for the item anchors, used in reverse scoring. Default: 4}
}
\description{
Score the Relationship and Sex Education (RSE) scale
}
\details{
Adds one column, \code{RSE_total}, to \code{df} containing the sum of scores.

Note: the default RSE scoring uses the mean of the items for the scales.

Note: the code assumes that pasting together the \code{item_prefix} and the numbers 1:10
will yield the 10 items from the test.
}
\author{
Michael Hallquist
}
