% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_rt.R
\name{score_rt}
\alias{score_rt}
\title{Score the Risk Taking -- 18 item scale (RT-18)}
\usage{
score_rt(
  df,
  item_prefix = "RT",
  max_impute = 0.2,
  drop_items = FALSE,
  input_codes = c(1, 2),
  min_value = 0,
  max_value = 1
)
}
\arguments{
\item{df}{a data.frame containing the 18 RT items to be scored}

\item{item_prefix}{a character prefix of the items names in \code{df} to be scored. Default: "RT"}

\item{max_impute}{the proportion of missingness [0..1) or number [1..n] of missing values per scale.
Below this threshold, the person subscale mean will be imputed for missing items.}

\item{drop_items}{whether to remove the item-level data from the \code{df}. Default: FALSE}

\item{input_codes}{the original T/F coding scheme of the input dataframe. These are converted to the specified min and max value.}

\item{min_value}{the minimum value for the item anchors. Default: 0}

\item{max_value}{the highest value for the item anchors. Default: 1}
}
\description{
Score the Risk Taking -- 18 item scale (RT-18)
}
\details{
Adds three columns, \code{RT_riskbehavior}, \code{RT_riskassessment}, and \code{RT_total}, to \code{df} containing
the different subscales, respectively.

Note: the default RT scoring uses the mean of the items for the scales.

Note: the code assumes that pasting together the \code{item_prefix} and the numbers 1:18
will yield the 18 items from the test.
}
\author{
Michael Hallquist
}
