% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_swls.R
\name{score_swls}
\alias{score_swls}
\title{Score the Satisfaction with Life Scale (Diener, Emmons, Larsen, & Griffin, 1985)}
\usage{
score_swls(df, item_prefix = "SWLS", max_impute = 0.2,
  drop_items = FALSE, min_value = 1, max_value = 7)
}
\arguments{
\item{df}{a data.frame containing the 5 SWLS items to be scored}

\item{item_prefix}{a character prefix of the items names in \code{df} to be scored. Default: "SWLS"}

\item{max_impute}{the proportion of missingness [0..1) or number [1..n] of missing values per scale.
Below this threshold, the person subscale mean will be imputed for missing items.}

\item{drop_items}{whether to remove the item-level data from the \code{df}. Default: FALSE}

\item{min_value}{the minimum value for the item anchors. Default: 1}

\item{max_value}{the highest value for the item anchors. Default: 7}
}
\description{
Score the Satisfaction with Life Scale (Diener, Emmons, Larsen, & Griffin, 1985)
}
\details{
Adds one column, \code{SWLS_total}, to \code{df} containing
the SWLS scale items.

Note: the default SWLS scoring uses the mean of the items for the scales.

Note: the code assumes that pasting together the \code{item_prefix} and the numbers 1:5
will yield the 5 items from the test.

See: \url{http://labs.psychology.illinois.edu/~ediener/SWLS.html}
}
\author{
Michael Hallquist
}
