% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_upps.R
\name{score_upps}
\alias{score_upps}
\title{Score the UPPS-P Impulsive Behavior scale -- revised 59 item version (Lynam, Smith, Whiteside, & Cyders, 2007)}
\usage{
score_upps(
  df,
  item_prefix = "UPPS_",
  max_impute = 0.2,
  drop_items = FALSE,
  keep_reverse_codes = FALSE,
  min_value = 1,
  max_value = 4,
  bad_items = NULL,
  add_alphas = TRUE
)
}
\arguments{
\item{df}{a data.frame containing the 59 UPPS items to be scored}

\item{item_prefix}{a character prefix of the items names in \code{df} to be scored. Default: "UPPS"}

\item{max_impute}{the proportion of missingness [0..1) or number [1..n] of missing values per scale.
Below this threshold, the person subscale mean will be imputed for missing items.}

\item{drop_items}{whether to remove the item-level data from the \code{df}. Default: FALSE}

\item{keep_reverse_codes}{whether to retain the reverse coded items (suffix "r")}

\item{min_value}{the minimum value for the item anchors, used in reverse scoring. Default: 1}

\item{max_value}{the highest value for the item anchors, used in reverse scoring. Default: 4}

\item{bad_items}{numeric value or vector of the items that need to be dropped before imputation or calculation of subscales}

\item{add_alphas}{whether to compute coefficient alpha for subscales and return a column attribute. Default: TRUE}
}
\description{
Score the UPPS-P Impulsive Behavior scale -- revised 59 item version (Lynam, Smith, Whiteside, & Cyders, 2007)
}
\details{
Adds five columns, \code{UPPS_negative_urgency}, \code{UPPS_lack_premeditation}, \code{UPPS_lack_perseverance}, \code{UPPS_sensation_seeking}, \code{UPPS_positive_urgency}, to \code{df} containing
the UPPS-P subscales, respectively.

Note: the default UPPS scoring uses the mean of the items for the scales.

Note: the code assumes that pasting together the \code{item_prefix} and the numbers 1:59
will yield the 59 items from the test.

See: \url{http://web.archive.org/web/20190710205448/http://www.impulsivity.org/measurement/UPPS-P_59_item.pdf}
}
\author{
Michael Hallquist, Zach Vig
}
