% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidp_report.R
\name{sidp_report}
\alias{sidp_report}
\title{Generate a Detailed Report for a Specific Subject from SIDP Survey Data}
\usage{
sidp_report(id)
}
\arguments{
\item{id}{Numeric or integer. The unique identifier of the subject for whom the report is to be generated.}
}
\value{
The function does not return a value but prints a detailed report to the console.
        The report includes general summaries and detailed breakdowns for each personality disorder.
}
\description{
This function generates a detailed report for a specific subject using SIDP survey data.
It utilizes the `import_score_neuromap_sidp` function to import and process the data,
then filters and summarizes the information for a given subject ID.
The report includes counts of symptoms by diagnosis, dimensional scores, and
DSM diagnostic codes for various personality disorders.
Detailed summaries for each personality disorder are also provided.
}
\examples{
# Example usage:
# sidp_report(710)


}
\author{
Nate Hall
}
