% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/deconvolve_funcs.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid transform}
\usage{
sigmoid(x, beta = 1)
}
\arguments{
\item{x}{value to be transformed by sigmoid}

\item{beta}{slope (steepness) of sigmoid transform}
}
\description{
Sigmoid transform

Sigmoid transform
}
\keyword{internal}
