% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_curation_transfer.R
\name{study_curation_transfer}
\alias{study_curation_transfer}
\title{Create data frame of messy directory structure}
\usage{
study_curation_transfer(dir)
}
\arguments{
\item{dir}{Name of the directory to be transferred.}
}
\description{
generate df for transfer to new study structure.
}
