% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_utility_fx.R
\name{visualizeDesignMatrix}
\alias{visualizeDesignMatrix}
\title{Visualize design matrix, including event onset times and run boundaries}
\usage{
visualizeDesignMatrix(d, outfile = NULL, runboundaries = NULL,
  events = NULL, includeBaseline = FALSE)
}
\description{
Visualize design matrix, including event onset times and run boundaries
}
