% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wait_for_job.R
\name{wait_for_job}
\alias{wait_for_job}
\title{This function pauses execution of an R script while a scheduled qsub job is not yet complete.}
\usage{
wait_for_job(jobid, sleep_interval = 10, max_wait = 60 * 60 * 24)
}
\arguments{
\item{jobid}{A valid id of an existing PBS job. (No validation on this argument at present)#'}

\item{sleep_interval}{How often to recheck the job status, in seconds. Default: 10}

\item{max_wait}{How long to wait on the job before giving up, in seconds. Default: 24 hours}
}
\value{
Nothing. Just returns when the blocking job completes.
}
\description{
It is intended to give you control over job dependencies within R when the formal PBS
depend approach is insufficient, especially in the case of a script that spawns child jobs that
need to be scheduled or complete before the parent script should continue.
}
\examples{
\dontrun{
wait_for_job('7968857.torque01.util.production.int.aci.ics.psu.edu')
}

}
\author{
Michael Hallquist
}
