% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve_funcs.R
\name{wgr_deconv_canonhrf_par}
\alias{wgr_deconv_canonhrf_par}
\title{R port of Wu et al. 2013, Medical Image Analysis
Adapted from the original provided by Daniele Marinazzo}
\usage{
wgr_deconv_canonhrf_par(data, thr = 1, event_lag_max, TR)
}
\arguments{
\item{data}{dataset to deconvolve}

\item{event_lag_max}{longest lag after an event that could generate a BOLD response}

\item{TR}{repetition time of scan (in seconds)}

\item{threshold}{(in SD units) for defining latent neural events}
}
\description{
R port of Wu et al. 2013, Medical Image Analysis
Adapted from the original provided by Daniele Marinazzo
}
\keyword{internal}
