% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_events}
\alias{example_events}
\title{example events structure for build_design_matrix}
\format{
An object of class \code{data.frame} with 200 rows and 6 columns.
}
\usage{
example_events
}
\description{
example events structure for build_design_matrix
}
\keyword{datasets}
