% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_signals}
\alias{example_signals}
\title{example signals structure for build_design_matrix}
\format{An object of class \code{list} of length 2.}
\usage{
example_signals
}
\description{
example signals structure for build_design_matrix
}
\keyword{datasets}
