% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmcmapply.R
\name{gmcmapply}
\alias{gmcmapply}
\title{expand.grid-based multi-core multivariate apply (mcmapply) wrapper for fast implementation of nested looping structures}
\usage{
gmcmapply(mvars, FUN, SIMPLIFY = FALSE, mc.cores = 1, ...)
}
\arguments{
\item{mvars}{a named list of variables to be combined and mapped over. These are equivalent to the layers of a nested for loop.}

\item{FUN}{user-defined function to apply over variables in mvars. N.B. This function is written to use the names of mvars as formal arguments to FUN. Thus, the majority of FUNs need not require any arguments to be specified. As long as the variable object names in FUN match the names of mvars,  gmapply will handle the translation of names(mvars) to FUN.}

\item{SIMPLIFY}{when set to TRUE, mimics the simplified output of Map to mapply.}

\item{mc.cores}{number of cores to utilize if running in parallel. Defaults to 1, which implements mapply.}

\item{...}{}
}
\value{
compiled returns of FUN
}
\description{
expand.grid-based multi-core multivariate apply (mcmapply) wrapper for fast implementation of nested looping structures
}
\examples{

\dontrun{
     # Example 1:
     # just make sure variables used in your function appear as the names of mvars
     myfunc <- function(...){
       return_me <- paste(l3, l1^2 + l2, sep = "_")
       return(return_me)
     }

     mvars <- list(l1 = 1:10,
                   l2 = 1:5,
                   l3 = letters[1:3])


     ### list output (mapply)
     lreturns <- gmcmapply(mvars, myfunc)

     ### concatenated output (Map)
     lreturns <- gmcmapply(mvars, myfunc, SIMPLIFY = TRUE)

     ## N.B. This is equivalent to running:
     lreturns <- c()
     for(l1 in 1:10){
       for(l2 in 1:5){
         for(l3 in letters[1:3]){
           lreturns <- c(lreturns,myfunc(l1,l2,l3))
         }
       }
     }

     ### concatenated outout run on 2 cores.
     lreturns <- gmcmapply(mvars, myfunc, SIMPLIFY = TRUE, mc.cores = 2)

    Example 2. Pass non-default args to FUN.
    ## Since the apply functions dont accept full calls as inputs (calls are internal), user can pass arguments to FUN through dots, which can overwrite a default option for FUN.

    ## update myfunc to have a default argument
     myfunc <- function(rep_letters = 3, ...){
       return_me <- paste(rep(l3, rep_letters), l1^2 + l2, sep = "_")
       return(return_me)
     }

     lreturns <- gmcmapply(mvars, myfunc, rep_letters = 1)

}

}
\author{
Nate Hall
}
