% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_utility_fx.R
\name{hrf_convolve_normalize}
\alias{hrf_convolve_normalize}
\title{This function convolves a regressor with a normalized HRF whose peak amplitude is 1.0}
\usage{
hrf_convolve_normalize(n_vols, times, durations, values, tr = 1,
  normeach = FALSE, rm_zeros = TRUE, center_values = TRUE,
  convmax_1 = FALSE, demean_convolved = FALSE, hrf_parameters = c(a1
  = 6, a2 = 12, b1 = 0.9, b2 = 0.9, cc = 0.35))
}
\arguments{
\item{n_vols}{The number of volumes (scans) to be output in the convolved regressor}

\item{times}{A vector of times (in seconds) specifying event onsets}

\item{durations}{A vector of durations (in seconds) for each event}

\item{values}{A vector of parametric values used as regressor heights prior to convolution}

\item{tr}{The repetition time in seconds}

\item{normeach}{Whether to normalize the HRF to 1.0 for each event separately. If TRUE, equivalent to dmUBLOCK(1).
If FALSE, the HRF is normed overall, equivalent to dmUBLOCK(0).}

\item{rm_zeros}{Whether to remove zeros from events vector prior to convolution. Generally a good idea since we typically center
values prior to convolution, and retaining zeros will lead them to be non-zero after mean centering.}

\item{center_values}{Whether to demean values vector before convolution. Default \code{TRUE}.}

\item{convmax_1}{Whether to rescale the convolved regressor to a maximum height of 1.}

\item{demean_convolved}{Whether to demean the regressor after convolution (default: \code{TRUE})}

\item{hrf_parameters.}{A named vector of parameters passed to \code{fmri.stimulus} that control the shape of the double gamma HRF.
Default: \code{c(a1 = 6, a2 = 12, b1 = 0.9, b2 = 0.9, cc = 0.35)}.}
}
\description{
It extends \code{fmri.stimulus} by allowing for two normalization approaches (building on AFNI dmUBLOCK):
  1) "evtmax_1": pre-convolution HRF max=1.0 normalization of each stimulus regardless of duration: identical to dmUBLOCK(1)
  2) "durmax_1": pre-convolution HRF max=1.0 normalization for long events (15+ sec) -- height of HRF is modulated by duration of event: identical to dmUBLOCK(0)
}
\author{
Michael Hallquist
}
\keyword{internal}
