% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_score_neuromap_self_reports.R
\name{import_score_neuromap_self_reports}
\alias{import_score_neuromap_self_reports}
\title{Import and Score NeuroMAP Self Report Data}
\usage{
import_score_neuromap_self_reports(
  split_output = TRUE,
  path = paste0(getwd(), "/neuromap_self_reports_repo")
)
}
\arguments{
\item{split_output}{if true, outputs a list object with "scores" and "items" containing score- and item- level data, respectively. Default: /code{TRUE}}

\item{path}{string containing path to house csv data exports. Creates a new directory if it doesn't already exist.
         Default creates a new folder called "neuromap_self_reports_repo" in the working directory.

Note: the embedded `import_neuromap_self_reports` exports new csvs into the data repo with a date/time stamp every time the function is run}
}
\description{
This is a comprehensive function for importing, scoring, and storing NeuroMAP self report data to be used in a data analysis pipeline.
It uses the `import_neuromap_self_reports` and applicable `score_` functions from the dependlab package.
}
\author{
Zach Vig
}
