% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_subIDs_names.R
\name{match_subIDs_names}
\alias{match_subIDs_names}
\title{Add subject IDs to data which only contain subject's names by matching names from tracker}
\usage{
match_subIDs_names(
  tracker_df,
  addto_df,
  col_first_name = "FIRSTNAME",
  col_last_name = "LASTNAME"
)
}
\arguments{
\item{tracker_df}{entire filepath to a downloaded copy of NeuroMAP Master UNC.xlsx}

\item{addto_df}{can be entire filepath to the downloaded coup of screener or if you want to add subjectIDs to the MasterData tab in the Master spreadsheet, simply mention "MasterData". default set to "MasterData".}

\item{col_first_name}{name of the column which contains the first name of the subject to match between the tracker and master/screener. default set to "FIRSTNAME" (column in MasterData).}

\item{col_last_name}{name of the column which contains the last name of the subject to match between the tracker and master/screener. default set to "LASTNAME" (column in MasterData).}
}
\value{
addto_df dataframe with ID column added in by matching names from study tracker
}
\description{
Some data only has subject names like the MasterData tab of the NeuroMAP Master spreadsheet and the Screener.
In order to map their screening psychopathology scores like AI, SPIN, PAI-BOR the subject IDs for S3 task analysis, 
use this script to match the name in the tracker and pull the corresponding subjectID from there.
}
\examples{
\dontrun{
  match_subIDs_names(tracker_df, master_df, col_first_name = "FIRSTNAME", col_last_name = "LASTNAME")
  match_subIDs_names(tracker_df, screener_df, col_first_name = "ContactForm_1", col_last_name = "ContactForm_2")
}

}
\author{
Nidhi Desai
}
