% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_fs.R
\name{score_fs}
\alias{score_fs}
\title{Score the Flourishing Scale (Diener, Wirtz, Tov, Kim-Prieto, Choi, Oishi, & Biswas-Diener, 2009)}
\usage{
score_fs(df, item_prefix = "FS", max_impute = 0.2,
  drop_items = FALSE, min_value = 1, max_value = 7)
}
\arguments{
\item{df}{a data.frame containing the 8 FS items to be scored}

\item{item_prefix}{a character prefix of the items names in \code{df} to be scored. Default: "FS"}

\item{max_impute}{the proportion of missingness [0..1) or number [1..n] of missing values per scale.
Below this threshold, the person subscale mean will be imputed for missing items.}

\item{drop_items}{whether to remove the item-level data from the \code{df}. Default: FALSE}

\item{min_value}{the minimum value for the item anchors. Default: 1}

\item{max_value}{the highest value for the item anchors. Default: 7}
}
\description{
Score the Flourishing Scale (Diener, Wirtz, Tov, Kim-Prieto, Choi, Oishi, & Biswas-Diener, 2009)
}
\details{
Adds one column, \code{FS_total}, to \code{df} containing
the FS scale items.

Note: the default FS scoring uses the mean of the items for the scales.

Note: the code assumes that pasting together the \code{item_prefix} and the numbers 1:8
will yield the 8 items from the test.

See: \url{http://labs.psychology.illinois.edu/~ediener/FS.html}
}
\author{
Michael Hallquist
}
