% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_iip90.R
\name{score_iip90}
\alias{score_iip90}
\title{Score the Inventory for Interpersonal Problems -- 90 items PD + circumplex version}
\usage{
score_iip90(
  df,
  item_prefix = "IIP",
  max_impute = 0.2,
  drop_items = FALSE,
  keep_octants = TRUE,
  min_value = 0,
  max_value = 4
)
}
\arguments{
\item{df}{a data.frame containing the 90 IIP items to be scored}

\item{item_prefix}{a character prefix of the items names in \code{df} to be scored. Default: "IIP"}

\item{max_impute}{the proportion of missingness [0..1] or number [1..n] of missing values per subscale.
Below this, the mean will be imputed for missing items.}

\item{drop_items}{whether to remove the item-level data from the \code{df}. Default: FALSE}

\item{keep_octants}{whether to retain the IIP octant scores. Default: FALSE}

\item{min_value}{the lowest value for the items anchors, used to check response validity. Default: 0}

\item{max_value}{the highest value for the items anchors, used to check response validity. Default: 4}
}
\description{
Score the Inventory for Interpersonal Problems -- 90 items PD + circumplex version
}
\details{
UPDATE ME!
Adds twenty columns, eight for octant scales, six for pd scales, and six for circumplex scales, to \code{df} containing
the octant, pd, and additional scales, respectively.

Note: the default ECR scoring uses the mean of the items for the scales.

Note: the code assumes that pasting together the \code{item_prefix} and the numbers 1:36
will yield the 36 items from the test.
}
\author{
Michael Hallquist
}
