% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_nmap_s2_self_report.R
\name{summarize_nmap_s2_self_report}
\alias{summarize_nmap_s2_self_report}
\title{Check Completion of NeuroMAP Self-Report Surveys}
\usage{
check_nmap_self_report_completion(...)
}
\value{
A data frame containing the `id` and the percentage of completion for various psychological measures.
}
\description{
This function retrieves self-report data from Qualtrics, processes it to calculate the percentage of completion for various psychological measures, and returns a summary data frame with the calculated completion percentages.
}
\details{
The function checks if the Qualtrics API key and base URL are set in the environment variables. If they are not set, the function issues a warning and exits. If the credentials are set, the function proceeds to:
\itemize{
  \item Import all self-report data from Qualtrics.
  \item Convert `Intro_ID` to a three-digit format.
  \item Rename `Intro_ID` to `id`.
  \item Remove specific columns related to attention checks and text fields.
  \item Arrange the data by `id` and select relevant columns.
  \item Calculate the percentage of completion for each psychological measure.
  \item Return a summary data frame with the calculated completion percentages.
}
}
\examples{
\dontrun{
# Set your Qualtrics API credentials
Sys.setenv(QUALTRICS_API_KEY = "your_api_key")
Sys.setenv(QUALTRICS_BASE_URL = "your_base_url")

# Check the self report completion
completion_data <- check_nmap_self_report_completion()
print(completion_data)
}

}
\author{
Rachel Velasquez
}
